+++ modules/spring-core/src/main/kotlin/com/esprito/spring/core/inspections/SpringBasePropertyInspection.kt
@@ -18,10 +18,12 @@ import com.esprito.spring.core.inspections.quickfix.ReplacementStringQuickFix
 import com.esprito.spring.core.inspections.utils.ResourceFileInspectionUtil
 import com.esprito.spring.core.service.SpringSearchService
 import com.esprito.spring.core.util.PropertyUtil
+import com.esprito.spring.core.util.PropertyUtil.isNotKebabCase
 import com.esprito.spring.core.util.PropertyUtil.propertyKey
 import com.esprito.spring.core.util.PropertyUtil.propertyKeyPsiElement
 import com.esprito.spring.core.util.PropertyUtil.propertyValue
 import com.esprito.spring.core.util.PropertyUtil.propertyValuePsiElement
+import com.esprito.spring.core.util.PropertyUtil.toKebabCase
 import com.esprito.spring.core.util.SpringCoreUtil
 import com.esprito.util.CacheKeyStore
 import com.intellij.codeInspection.*
@@ -120,7 +122,13 @@ abstract class SpringBasePropertyInspection : SpringBaseLocalInspectionTool() {
                 )
             }
 
-            val foundProperties = properties.filter { PropertyUtil.isSameProperty(it.name, fileProperty.key) }
+            val key = fileProperty.key
+            if (isNotKebabCase(key)) {
+                problems +=
+                    keyShouldBeKebabProblemDescriptor(manager, psiKey, isOnTheFly, key)
+            }
+
+            val foundProperties = properties.filter { PropertyUtil.isSameProperty(it.name, key) }
             val placeholders = DefinedConfigurationPropertiesSearch.getInstance(module.project)
                 .getAllPlaceholders(module)
 
@@ -128,13 +136,12 @@ abstract class SpringBasePropertyInspection : SpringBaseLocalInspectionTool() {
                 && !isPropertyMapKey(fileProperty, properties)
                 && !isPropertyListKey(fileProperty, properties)
                 && placeholders.none {
-                    PropertyUtil.isSameProperty(fileProperty.key, it)
+                    PropertyUtil.isSameProperty(key, it)
                 }
             ) {
                 val psiReferences = SpringSearchService.getInstance(elementFileProperty.project)
                     .getAllReferencesToElement(elementFileProperty)
                 if (psiReferences.isEmpty()) {
-                    val key = fileProperty.key
                     problems += manager.createProblemDescriptor(
                         psiKey,
                         SpringCoreBundle.message("esprito.spring.inspection.properties.key.unresolved", key),
@@ -160,6 +167,13 @@ abstract class SpringBasePropertyInspection : SpringBaseLocalInspectionTool() {
         }
         return problems
     }
+
+    abstract fun keyShouldBeKebabProblemDescriptor(
+        manager: InspectionManager,
+        psiKey: PsiElement,
+        isOnTheFly: Boolean,
+        key: String
+    ): ProblemDescriptor
 
     private fun getProblemPropertyDeprecated(
         manager: InspectionManager,
@@ -220,7 +234,7 @@ abstract class SpringBasePropertyInspection : SpringBaseLocalInspectionTool() {
             problems += PropertyUtil.getPlaceholders(value) { placeholder, range ->
                 PlaceholderWithRange(placeholder, range)
             }.asSequence()
-                .filter { isNotSnakeCase(it.placeholder) }
+                .filter { isNotKebabCase(it.placeholder) }
                 .mapTo(mutableListOf()) {
                     manager.createProblemDescriptor(
                         psiValue,
@@ -228,7 +242,7 @@ abstract class SpringBasePropertyInspection : SpringBaseLocalInspectionTool() {
                         SpringCoreBundle.message("esprito.spring.inspection.properties.value.should.be.kebab"),
                         ProblemHighlightType.WARNING,
                         isOnTheFly,
-                        ReplacementStringQuickFix(it.placeholder, toSnakeCase(it.placeholder), psiValue, it.range)
+                        ReplacementStringQuickFix(it.placeholder, toKebabCase(it.placeholder), psiValue, it.range)
                     )
                 }
         }
@@ -276,29 +290,6 @@ abstract class SpringBasePropertyInspection : SpringBaseLocalInspectionTool() {
             }
         }
         return problems
-    }
-
-    private fun isNotSnakeCase(placeholder: String): Boolean {
-        return placeholder.any { it.isUpperCase() || it == '_' }
-    }
-
-    private fun toSnakeCase(placeholder: String): String {
-        if (placeholder.isEmpty()) return ""
-
-        var prev: Char? = null
-        val stringBuilder = StringBuilder()
-        for (current in placeholder) {
-            if (current.isUpperCase()) {
-                if (prev != null && prev.isLowerCase()) {
-                    stringBuilder.append('_')
-                }
-            }
-            stringBuilder.append(current.lowercaseChar())
-
-            prev = current
-        }
-
-        return stringBuilder.replace(UNDERSCORES_REGEX, "-")
     }
 
     data class PlaceholderWithRange(val placeholder: String, val range: TextRange)
@@ -757,7 +748,6 @@ abstract class SpringBasePropertyInspection : SpringBaseLocalInspectionTool() {
         val PROHIBITED_IN_PROFILE_PROPERTIES =
             setOf("spring.profiles.include", "spring.profiles.active", "spring.profiles.default")
         val PROFILE_PROPERTIES_FILE_MASK = Regex("application-.*\\.(properties|yml|yaml)")
-        val UNDERSCORES_REGEX = Regex("_+")
     }
 
 }