+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed: support CoroutineCrudRepository
 
 ### Spring Core/Boot
+- Inspection: `Should be kebab-case` for property key
 - Fixed: kebab-case inspection description
 - Fixed: Yaml PropertyLineMarker only for leaf elements
 - Inspection: @EntityScan package support